backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic[email]
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
databases[sqlite]==0.8.0
sqlalchemy==1.4.42
aiosqlite==0.19.0
bcrypt==4.1.1
python-dotenv==1.0.0
asyncpg==0.29.0
redis==5.0.1
celery==5.3.6
minio==7.2.3
qdrant-client==1.7.0 